+++ src/main.ts
@@ -73,8 +73,13 @@ const main = async () => {
       stack.up({ onOutput, ...config.options }).then((r) => r.stdout),
     refresh: () =>
       stack.refresh({ onOutput, ...config.options }).then((r) => r.stdout),
-    destroy: () =>
-      stack.destroy({ onOutput, ...config.options }).then((r) => r.stdout),
+    destroy: async () => {
+      const { stdout, stderr } = await stack.destroy({ onOutput, ...config.options });
+      if (config.deleteStack) {
+        await stack.workspace.removeStack(stack.name);
+      }
+      return stdout;
+    },
     preview: async () => {
       const { stdout, stderr } = await stack.preview(config.options);
       onOutput(stdout);